requirements/base.txt
Django<1.9

# Configuration
django-environ==0.4.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.24.1

# PostgreSQL adapter for the Python
psycopg2==2.6

# Images
Pillow==2.8.1

# Rest Framework
djangorestframework==3.4.1

# django-filter==0.7
django-filter==0.10.0

# static mgt
django-compressor==2.0

# internacionalization admin interface
django-rosetta==0.7.6

jsonfield==1.0.3
django-suit==0.2.13

docutils==0.12

# Pygments==1.6
Pygments==2.0.2

# django-autoslug==1.7.2
django-autoslug==1.9.3

# metron==1.3.1
# metron==1.3.5
pinax-webanalytics==2.0.4

django-braces==1.3.1
# django-braces==1.8.0

# unicodecsv==0.9.4
unicodecsv==0.13.0

raven==5.5.0

# Markdown==2.4.1
Markdown==2.6.2
django-markdown==0.8.4
twitter==1.17.1

django-extensions==1.5.5

# Themes
git+https://github.com/institutotim/timtec-theme-timtec.git#master
git+https://github.com/institutotim/timtec-theme-if.git#master
git+https://github.com/lafabbricabr/timtec-theme-nutrir.git#master

# JWT
PyJWT==1.5.2

# Selenium
selenium==3.0.2

python-docx==0.8.6
